change(middle-end) - use simple type alias instead of custom nested data type for Char and Bool
Even though its nice to have a specific WackBool and WackChar to handle
errors and handle bugs. However these are the jobs of the frontend to
ensure. Either way, you'd have to adjust both frontend and middle-end to
change things. In WackChar, I think its fine to overcomplicate but
WackBool is overkill especially with representing bools as u8.
Regardless in the case of a compiler: when doing optimizations reaching
1 level extra is complicated only to put them back.
Plus we still have to care about low level details considering we're
following the wacc spec / have to care about low level details here